Fixed unoperable GUI when joining a server
