+++ TODOs.txt
@@ -111,6 +111,12 @@ TODO: BUG: Crowbar is displayed in the slot for erasing items at Creative GUI.
 DONE: BUG: Right shift+clicking an item in inventory doesn't redraw GUI. 
 
 ------------
+Server dev
+------------
+TODO: Problems with butter churn not working with fluid bottles
+TODO: Fire symbol blinking for heated devices. Heated devices not working.
+
+------------
 Mod Integration: Pam's Harvestcraft
 ------------
 TODO: Add recipes with HC honey combs
