+++ src/main/java/growthcraft/cellar/client/ClientProxy.java
@@ -5,13 +5,10 @@ import growthcraft.cellar.client.gui.GuiBrewKettle;
 import growthcraft.cellar.client.gui.GuiCultureJar;
 import growthcraft.cellar.client.gui.GuiFermentBarrel;
 import growthcraft.cellar.client.gui.GuiFruitPress;
-import growthcraft.cellar.client.gui.GuiHandler;
 import growthcraft.cellar.common.CommonProxy;
 import growthcraft.cellar.common.Init;
-import growthcraft.cellar.shared.Reference;
 import growthcraft.cellar.shared.init.GrowthcraftCellarItems;
 import growthcraft.cellar.shared.init.GrowthcraftCellarItems.EnumYeast;
-import net.minecraftforge.fml.common.network.NetworkRegistry;
 
 public class ClientProxy extends CommonProxy {
     @Override
@@ -28,10 +25,7 @@ public class ClientProxy extends CommonProxy {
     
     @Override
     public void init() {
-    	super.init();
-    	GrowthcraftCellar.guiProvider = new GuiHandler();
-    	NetworkRegistry.INSTANCE.registerGuiHandler(Reference.MODID, GrowthcraftCellar.guiProvider);
-    	
+    	super.init();    	
     	registerSpecialRenders();
     	GrowthcraftCellar.guiProvider.register("growthcraft_cellar:fruit_press", GuiFruitPress.class);
     	GrowthcraftCellar.guiProvider.register("growthcraft_cellar:brew_kettle", GuiBrewKettle.class);
