+++ src/main/java/growthcraft/cellar/client/gui/GuiHandler.java
@@ -1,11 +1,12 @@
 package growthcraft.cellar.client.gui;
 
-import growthcraft.core.shared.client.gui.GrowthcraftGuiProvider;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.common.network.IGuiHandler;
 
 import javax.annotation.Nullable;
+
+import growthcraft.core.shared.inventory.GrowthcraftGuiProvider;
 
 public class GuiHandler extends GrowthcraftGuiProvider implements IGuiHandler {
